feat: 更新微调代码与文档，支持新版本 transformers 和 QLoRA 训练流程 🚀
1. **merge_lora.py**:
   - 将 `torch_dtype` 参数替换为 `dtype`，以兼容 transformers 4.45+ 版本的更新。🛠️
   - 修改了加载基础模型的相关代码，确保在新版本中正常运行。

2. **train_qlora.py**:
   - 同样将 `torch_dtype` 替换为 `dtype`，并添加了对旧配置的兼容处理。🔄
   - 更新了 SFTConfig 的参数，使用 `max_length` 替代 `max_seq_length`，以适应新版本的要求。📏
   - 增加了对 `assistant_only_loss` 的条件判断，确保在 YAML 配置中明确开启时才传递该参数。✅

3. **PPT_Report.md**:
   - 新增了关于基于大语言模型构建免疫细胞知识图谱的微调阶段汇报文档，详细描述了项目目标、已有数据、训练数据构造、分层采样策略、训练样本格式、模型与训练框架选型等内容。📊
   - 包含了训练超参数、硬件资源配置、实验监控等信息，便于后续参考与复现。📋

4. **merge_lora.ipynb**:
   - 新增了合并 LoRA Adapter 的 Notebook 版本，提供交互式操作界面，便于用户在训练后合并模型权重。📓

5. **train_qlora.ipynb**:
   - 新增了 QLoRA 微调的 Notebook 版本，便于在服务器上进行交互式调试，确保训练过程的灵活性与可视化。🖥️

这些更改旨在提升代码的兼容性与可读性，同时为用户提供更友好的操作界面，确保在新版本的环境中顺利运行。✨

+++ src/Fine_tuning/training/merge_lora.py
@@ -110,7 +110,7 @@ def main() -> None:
     print(f"[加载] 基础模型：{base_model}  dtype={args.torch_dtype}  device_map={device_map}")
     base = AutoModelForCausalLM.from_pretrained(
         base_model,
-        torch_dtype=dtype,
+        dtype=dtype,                                # transformers 4.45+ 使用 dtype 替代 torch_dtype
         device_map=device_map,
         trust_remote_code=cfg["model"].get("trust_remote_code", True),
     )
